requirements_simple.txt
# Simplified requirements - NO DATABASE DEPENDENCIES
# For Azure Web App deployment

# Core MCP and web framework
fastmcp>=0.1.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=20.1.0

# Power BI authentication  
msal>=1.24.0
requests>=2.31.0

# Security
pyjwt>=2.8.0

# Basic utilities (no database packages)
python-dotenv>=1.0.0